third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-hover-state-iframe-expected.txt

Gestures go here

Tests that the hover effect is cleared by a gesture tap between iframe and document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Verify the hover is not initially set
PASS getHoverActiveState(box1) is "default"
PASS getHoverActiveState(box2) is "default"
Verify the hover effect is set in iframe
PASS getHoverActiveState(box1) is "default"
PASS getHoverActiveState(box2) is "hovered"
Verify the hover effect is cleared in iframe
PASS getHoverActiveState(box1) is "hovered"
PASS getHoverActiveState(box2) is "default"
PASS successfullyParsed is true

TEST COMPLETE
